Change e6809 to sized integers

+++ e6809.h
@@ -3,10 +3,10 @@
 
 /* user defined read and write functions */
 
-extern uint8_t (*e6809_read8) (uint16_t address);
-extern void (*e6809_write8) (uint16_t address, uint8_t data);
+extern uint16_t (*e6809_read8) (uint16_t address);
+extern void (*e6809_write8) (uint16_t address, uint16_t data);
 
 void e6809_reset (void);
-unsigned e6809_sstep (unsigned irq_i, unsigned irq_f);
+uint16_t e6809_sstep (uint16_t irq_i, uint16_t irq_f);
 
 #endif
